搜索详情页完成

+++ public/front/js/search.js
@@ -12,8 +12,8 @@
     //添加功能
     $('.lt_search button').on('click',function(){
         var value = $('.lt_search input').val().trim();
-        //如果为空提示信息
         $(".lt_search input").val('');
+        //如果为空提示信息
         if(value == "") {
             mui.toast("请输入搜索关键字");
             return;
@@ -30,14 +30,20 @@
         arr.unshift(value);
         localStorage.setItem('search_list',JSON.stringify(arr))
         render()
+
     })
 
     //删除功能
     $('.lt_history').on('click','.btn_delete',function(){
-        var arr = getStorage();
-        arr.splice($(this).data('index'),1)
-        localStorage.setItem('search_list',JSON.stringify(arr))
-        render()
+        that=$(this)
+        mui.confirm("你确定要删除这条历史记录吗?", "温馨提示", ["取消", "确认"], function (e) {
+            if (e.index === 1) {
+                var arr = getStorage();
+                arr.splice(that.data('index'), 1)
+                localStorage.setItem('search_list', JSON.stringify(arr))
+                render()
+            }
+        })
     })
 
     //清空记录功能
@@ -51,11 +57,4 @@
         })
 
     })
-
-    //获得缓存记录
-    function getStorage(){
-        var history=localStorage.getItem('search_list')||"[]";
-        arr= JSON.parse(history);
-        return arr
-    }
 })
